requirements.txt
python-dotenv==1.1.0
langchain==0.3.25
langchain-core==0.3.59
langgraph==0.4.3
pydantic==2.11.4
typing-extensions==4.13.2
openai==1.78.0
langchain-openai==0.3.16
langmem==0.0.25
numpy==2.2.5
langchain-anthropic==0.3.5
langchain-groq==0.2.3
langchain-deepseek==0.1.2
pandas==2.2.3
matplotlib==3.9.2
tabulate==0.9.0
colorama==0.4.6
questionary==2.1.0
rich==13.9.4
langchain-google-genai==2.0.11
fastmcp==2.3.3
pytest==8.3.5
black==23.7.0
isort==5.12.0
flake8==6.1.0 
aiohttp-rpc==1.3.3
lightgbm==4.6.0
uvicorn==0.34.2
fastapi==0.115.12
python-multipart==0.0.20
mcp-server==0.1.4
aiohttp==3.9.3
websockets>=10.0 
requests>=2.31.0
httpx==0.28.1
pymongo==4.13.0
sseclient==0.0.27
websocket-client==1.8.0